+++ app/services/content-parser.ts
@@ -1,17 +1,16 @@
 import Service, { service } from '@ember/service';
 import { htmlSafe } from '@ember/template';
-// import yabbcode from 'ya-bbcode';
-import yabbcode from './bbcode/ya-bbcode';
 import { emojis } from 'potber-client/utils/icons';
 import MessagesService from './messages';
-import { parseMod } from './content-parser/mod';
-import { parseTex } from './content-parser/tex';
 import { parseVideo } from './content-parser/video';
 import { parseUrl } from './content-parser/url';
+import { parseSimpleTags } from './content-parser/simple-tags';
+import { parseImg } from './content-parser/img';
+import { parseQuote } from './content-parser/quote';
+import { parseList } from './content-parser/list';
 
 export default class ContentParserService extends Service {
   @service declare messages: MessagesService;
-  parser = this.registerCustomTags(new yabbcode());
 
   /**
    * Parses post content to HTML and returns the result.
@@ -20,17 +19,15 @@ export default class ContentParserService extends Service {
    */
   parsePostContent(input: string) {
     let output = input;
-    // Simple formatting tags
-    output = parseMod(output);
-    output = parseTex(output);
+    output = parseSimpleTags(output);
+    output = parseImg(output);
     output = parseVideo(output, window.location);
     output = parseUrl(output);
+    output = parseQuote(output, window.location);
+    output = parseList(output);
     output = this.parseTable(output);
-    output = this.parseList(output);
-    output = this.parseQuoteUsernames(output);
-    output = this.parser.parse(output);
     output = this.parseEmojis(output);
-    output = this.cleanup(output);
+    output = this.format(output);
     return htmlSafe(output);
   }
 
@@ -53,86 +50,10 @@ export default class ContentParserService extends Service {
   }
 
   /**
-   * Parses [list] tags. Does not support tag nesting.
+   * Parses emojis.
    * @param input The input string.
    * @returns The output string.
    */
-  parseList(input: string) {
-    const regex = new RegExp(/(?:(\[list\])(.|\n)*?(\[\/list\]))/g);
-    if (!regex.test(input)) return input;
-    let output = input;
-    const matches = input.match(regex);
-    if (matches) {
-      for (const match of matches) {
-        let replacement = match;
-        replacement = replacement.replace(/\[list\]/, '<ul>');
-        const listItemMatches = replacement.match(/\[\*\]/g);
-        if (!listItemMatches) {
-          // If the list does not contain any list items, add ending tag right away and return
-          replacement = replacement.replace(/\[\/list\]/, '</ul>');
-          output.replace(match, replacement);
-          break;
-        } else {
-          for (const i in listItemMatches) {
-            if (i === '0') {
-              // Add no ending tag before first list item
-              replacement = replacement.replace(
-                listItemMatches[i] as string,
-                '<li>'
-              );
-            } else {
-              // Add ending tags before all other list items
-              replacement = replacement.replace(
-                listItemMatches[i] as string,
-                '</li><li>'
-              );
-            }
-          }
-        }
-        replacement = replacement.replace(/\[\/list\]/, '</li></ul>');
-        output = output.replace(match, replacement);
-      }
-    }
-    return output;
-  }
-
-  /**
-   * Registers custom tags for yabbcode parser. Default tags can be found here:
-   * https://github.com/nodecraft/ya-bbcode/blob/main/ya-bbcode.js
-   * The parser is documented at https://github.com/nodecraft/ya-bbcode.
-   * @param parser The yabbcode parser instance.
-   * @returns The modified yabbcode parser instance.
-   */
-  private registerCustomTags(parser: yabbcode) {
-    parser.registerTag('video', {
-      type: 'ignore',
-    });
-    parser.registerTag('table', {
-      type: 'ignore',
-    });
-    parser.registerTag('list', {
-      type: 'ignore',
-    });
-    parser.registerTag('url', {
-      type: 'ignore',
-    });
-    parser.registerTag('quote', {
-      type: 'content',
-      replace: this.parseQuote,
-    });
-    parser.registerTag('s', {
-      type: 'replace',
-      open: '<s>',
-      close: '</s>',
-    });
-    parser.registerTag('spoiler', {
-      type: 'replace',
-      open: '<label class="spoiler"><input class="spoiler-input" type="checkbox"/><p class="spoiler-header">👀 Spoiler anzeigen</p><span class="spoiler-content">',
-      close: '</span></label>',
-    });
-    return parser;
-  }
-
   private parseEmojis(input: string) {
     let output = input;
     for (const emoji of emojis) {
@@ -145,47 +66,13 @@ export default class ContentParserService extends Service {
   }
 
   /**
-   * Quotes can contain usernames which again can contain square brackets, which
-   * will make yabbcode misinterpret the end of the opening tag. We will need
-   * to prepare those so that yabbcode will have a easier time parsing them.
+   * Formats the content and parses line breaks.
+   * @param input The input string.
+   * @returns The output string.
    */
-  private parseQuoteUsernames(input: string) {
-    // eslint-disable-next-line no-useless-escape
-    const quoteAttributesRegex = RegExp(/(?:(\")(.*)(\"))/g);
-    const matches = input.match(quoteAttributesRegex);
-    if (!matches || matches.length === 0) return input;
+  format(input: string): string {
     let output = input;
-    for (const match of matches) {
-      const replacement = match.replace('[', '<SQBO>').replace(']', '<SQBC>');
-      output = output.replace(match, replacement);
-    }
-    return output;
-  }
-
-  private parseQuote(attr: string, content: string) {
-    if (!attr) {
-      return `<span class="quote"><blockquote>${content}</blockquote></span>`;
-    }
-    const userNameMatches = attr.match(/(?:[\\"](.*)[\\"])/);
-    let userName = '';
-    if (userNameMatches && userNameMatches.length > 1) {
-      userName = userNameMatches[1] as string;
-    } else {
-      userName = 'Unknown';
-    }
-    const ids = attr.replace(/[^0-9,]/g, '').split(',');
-    if (ids.length >= 2) {
-      const threadId = ids[0];
-      const postId = ids[1];
-      const url = `${window.location.protocol}//${window.location.host}/thread?TID=${threadId}&PID=${postId}#reply_${postId}`;
-      return `<span class="quote"><a class="quote-header" href="${url}"><p>${userName}</p></a><blockquote>${content}</blockquote></span>`;
-    }
-    return `<span class="quote"><span class="quote-header"><p>${userName}</p></span><blockquote>${content}</blockquote></span>`;
-  }
-
-  cleanup(input: string) {
-    let output = input;
-    output = output.replaceAll('<SQBO>', '[').replaceAll('<SQBC>', ']');
+    output = output.replaceAll(/\n/g, '<br/>');
     return output;
   }
 }