+++ tests/unit/services/content-parser/url-test.ts
@@ -3,7 +3,7 @@ import { module, test } from 'qunit';
 import { urlTagMocks } from './_mock/url';
 import { parseUrl } from 'potber-client/services/content-parser/url';
 
-module('Unit | Service | ContentParser', (hooks) => {
+module('Unit | Service | ContentParser | [url]', (hooks) => {
   setupTest(hooks);
 
   test('Parses all [url] tags.', (assert) => {